requirements.txt
numpy==1.24.4
pandas==2.0.3
Pillow==9.5.0
Pillow==10.0.0
pydeck==0.8.0
pytest==7.4.0
python-dotenv==1.0.0
Requests==2.31.0
streamlit==1.25.0
pytest-cov
